+++ views/authors/author.view.php
@@ -1,0 +1,7 @@
+<?php component("head") ?>
+
+<main>
+    <?= $author['name'] ?>
+</main>
+
+<?php component("footer") ?>