+++ views/components/nav.php
@@ -47,7 +47,7 @@ use App\Core\Auth;
               </form>
             </li>
             <li>
-              <a href="/my-posts" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 dark:hover:bg-gray-600 dark:text-gray-200 dark:hover:text-white">My Posts</a>
+              <a href="/authors/posts" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 dark:hover:bg-gray-600 dark:text-gray-200 dark:hover:text-white">My Posts</a>
             </li>
           </ul>
         </div>
@@ -80,6 +80,9 @@ use App\Core\Auth;
         </li>
         <li>
           <a href="/modules" class="<?= isUrl('/modules') ? 'text-blue-600' : 'text-gray-900' ?> block py-2 pl-3 pr-4 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700" <?= isUrl('/modules') ? "aria-current='page'" : null ?>>Modules</a>
+        </li>
+        <li>
+          <a href="/authors" class="<?= isUrl('/authors') ? 'text-blue-600' : 'text-gray-900' ?> block py-2 pl-3 pr-4 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700" <?= isUrl('/authors') ? "aria-current='page'" : null ?>>Authors</a>
         </li>
         <li>
           <a href="/contact" class="<?= isUrl('/contact') ? 'text-blue-600' : 'text-gray-900' ?> block py-2 pl-3 pr-4 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700" <?= isUrl('/contact') ? "aria-current='page'" : null ?>>Contact</a>